#3696 [YSQL] Flush buffered operations before calculating query time
Summary:
Query duration is calculated in `standard_ExecutorFinish` and `pgss_ProcessUtility` functions.
To calculate duration correctly all buffered operations must be flushed before that time.
Now flush is performed in `standard_ExecutorFinish` and hook of `ProcessUtility` functions.
`ProcessUtility` can call `standard_ExecutorFinish` internally.
To avoid multiple flushing while processing single statement next helper functions are introduced:
-`YBBeginOperationsBuffering`
-`YBEndOperationsBuffering`
-`YBResetOperationsBuffering`
These function uses internal static counter `buffering_nesting_level` and calls `YBCPgStartOperationsBuffering`/`YBCPgFlushBufferedOperations` only once.

Test Plan: ./yb_build.sh --java-test org.yb.pgsql.TestYSQLMetrics

Reviewers: neha, mihnea

Reviewed By: mihnea

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D8125

+++ java/yb-pgsql/src/test/java/org/yb/pgsql/TestPgTransactions.java
@@ -696,7 +696,7 @@ public class TestPgTransactions extends BasePgSQLTest {
                              "UPDATE test SET v = 3 WHERE k = 1", 1);
 
     // Verify JDBC single-row prepared statements use non-txn path.
-    int oldTxnValue = getMetricCounter(TRANSACTIONS_METRIC);
+    long oldTxnValue = getMetricCounter(TRANSACTIONS_METRIC);
 
     PreparedStatement insertStatement =
       connection.prepareStatement("INSERT INTO test VALUES (?, ?)");
@@ -715,7 +715,7 @@ public class TestPgTransactions extends BasePgSQLTest {
     updateStatement.setInt(2, 1);
     updateStatement.executeUpdate();
 
-    int newTxnValue = getMetricCounter(TRANSACTIONS_METRIC);
+    long newTxnValue = getMetricCounter(TRANSACTIONS_METRIC);
     assertEquals(oldTxnValue, newTxnValue);
   }
 }
